refactor: rewrite rollup config with ts & swc
